Saving Progress. Added Create Rubric button to the top right of the screen. The page also allows you to select a category but is currently not showing the category name
