+++ FrontEndReact/src/View/Admin/Add/AddCustomRubric/AddCustomRubric.js
@@ -1,97 +1,157 @@
-import React from "react";
-import CustomDataTable from "../../../Components/CustomDataTable";
-import { IconButton, TextField } from "@mui/material";
-import VisibilityIcon from "@mui/icons-material/Visibility";
-import AddCircleIcon from "@mui/icons-material/AddCircle";
+import React, { useState } from "react";
+import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Checkbox, Card, Collapse, IconButton } from "@mui/material";
+import { KeyboardArrowDown, KeyboardArrowUp } from "@mui/icons-material";
 import RemoveCircleIcon from "@mui/icons-material/RemoveCircle";
-import { genericResourcePOST, parseCategoriesByRubrics } from "../../../../utility";
-import Cookies from "universal-cookie";
-import ErrorMessage from "../../../Error/ErrorMessage";
+import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { Grid } from "@mui/material";
+import ErrorMessage from "../../../Error/ErrorMessage";
+import CustomDataTable from "../../../Components/CustomDataTable";
+import CustomButton from "./Components/CustomButton";
 
+// NOTE: Custom Theme for the Collapsible table
+const customTheme = createTheme({
+  components: {
+    // NOTE: Style for the table head
+    MuiTableCell: { 
+      styleOverrides: {
+        head: { 
+          backgroundColor: "#A4C4F4", 
+          padding: "23px",
+        },
+        root: {
+          padding: "19px",
+        },
+      },
+    },
+    MuiButton: {
+      styleOverrides: {
+        root: {
+          fontSize: "14px",
+        },
+      },
+    },
+    MuiToolbar: {
+      styleOverrides: {
+        root: {
+          display: "flex",
+          alignItems: "baseline",
+          padding: "0.5rem",
+        },
+      },
+    },
+  },
+});
 
+const CollapsableTableComponent = ({ categories, rubrics, onCategorySelect }) => {
+
+  // NOTE: Manage whether the rubric was clicked or not
+  const [openRubric, setOpenRubric] = useState(null);
+
+  const handleRubricClick = (rubric_id) => {
+    setOpenRubric(openRubric === rubric_id ? null : rubric_id);
+  };
+
+  // NOTE: Manage whether the category was clicked or not
+  const [checkedCategories, setCheckedCategories] = useState([]);
+
+  const handleCheckboxChange = (category_id) => {
+    const isChecked = checkedCategories.includes(category_id);
+
+    if (isChecked) {
+      setCheckedCategories(
+        checkedCategories.filter((id) => id !== category_id),
+      );
+
+      // Call onCategorySelect with isSelected set to false
+      onCategorySelect(category_id, false);
+    } else {
+      setCheckedCategories([...checkedCategories, category_id]);
+
+      // Call onCategorySelect with isSelected set to true
+      onCategorySelect(category_id, true);
+    }
+  };
+
+  return (
+    <ThemeProvider theme={customTheme}>
+      <TableContainer component={Card}>
+        <Table>
+          <TableHead>
+            <TableRow>
+              <TableCell>Rubric</TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {rubrics.map((rubric) => (
+              <React.Fragment key={rubric.rubric_id}>
+                <TableRow onClick={() => handleRubricClick(rubric.rubric_id)}>
+                  <TableCell>
+                    {rubric.rubric_name}
+                    {openRubric === rubric.rubric_id ? (
+                      <KeyboardArrowUp />
+                    ) : (
+                        <KeyboardArrowDown />
+                      )}
+                  </TableCell>
+                </TableRow>
+                <TableRow>
+                  <TableCell
+                    style={{ paddingBottom: 0, paddingTop: 0 }}
+                    colSpan={6}
+                  >
+                    <Collapse
+                      in={openRubric === rubric.rubric_id}
+                      timeout="auto"
+                      unmountOnExit
+                    >
+                      <Table>
+                        <TableBody>
+                          {categories
+                            .filter(
+                              (category) =>
+                                category.rubric_id === rubric.rubric_id,
+                            )
+                            .map((category) => (
+                              <TableRow key={category.category_id}>
+                                <TableCell component="th" scope="row">
+                                  <Checkbox
+                                    checked={checkedCategories.includes(
+                                      category.category_id,
+                                    )}
+                                    onChange={() =>
+                                      handleCheckboxChange(category.category_id)
+                                    }
+                                  />
+                                  {category.category_name}
+                                </TableCell>
+                              </TableRow>
+                            ))}
+                        </TableBody>
+                      </Table>
+                    </Collapse>
+                  </TableCell>
+                </TableRow>
+              </React.Fragment>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
+
+    </ThemeProvider>
+  );
+};
 class AddCustomRubric extends React.Component {
   constructor(props) {
     super(props);
 
     this.state = {
       chosen_rubric: null,
-      selected_categories: this.props.chosen_category_json,
-      category_map: this.props.category_map,
+      selected_categories: {},
+      category_map: {},
       isLoaded: null,
       errorMessage: null,
-    }
-    this.add_category = (category_id) => {
-      var updated_categories = this.state.selected_categories;
-
-      updated_categories[category_id] = true;
-
-      this.setState({
-        selected_categories: updated_categories
-      });
-    }
-
-    this.remove_category = (category_id) => {
-      var updated_categories = this.state.selected_categories;
-
-      updated_categories[category_id] = false;
-
-      this.setState({
-        selected_categories: updated_categories
-      });
-    }
+    };
   }
-
-  handleCreateRubric = (picked_categories) => {
-    var category_ids = [];
-
-    for (var category_index = 0; category_index < picked_categories.length; category_index++) {
-      category_ids = [...category_ids, picked_categories[category_index]["category_id"]];
-    }
-
-    if (document.getElementById("rubric_name_input").value === "") {
-      this.setState({
-        isLoaded: true,
-        errorMessage: "Missing Rubric Name."
-      });
-
-      return;
-    }
-
-    if (document.getElementById("rubric_description_input").value === "") {
-      this.setState({
-        isLoaded: true,
-        errorMessage: "Missing Rubric Description."
-      });
-
-      return;
-    }
-
-    if (category_ids.length === 0) {
-      this.setState({
-        isLoaded: true,
-        errorMessage: "Missing categories, at least one category must be selected."
-      });
-
-      return;
-    }
-
-    var cookies = new Cookies();
-
-    genericResourcePOST(
-      `/rubric`,
-      this,
-      JSON.stringify({
-        rubric: {
-          rubric_name: document.getElementById("rubric_name_input").value,
-          rubric_description: document.getElementById("rubric_description_input").value,
-          owner: cookies.get('user')['user_id'],
-        },
-
-        categories: category_ids,
-      })
-    );
-  };
 
   componentDidUpdate() {
     if (this.state.isLoaded === true && this.state.errorMessage === null) {
@@ -99,133 +159,21 @@ class AddCustomRubric extends React.Component {
     }
   }
 
-  render() {
-    var rubrics = this.props.rubrics;
-    var categories = this.props.categories;
-    var categories_by_rubric_id = parseCategoriesByRubrics(rubrics, categories);
+  handleCategorySelect = (category_id, isSelected) => {
+    const selectedCategories = {...this.state.selected_categories };
+    if (isSelected) {
+      selectedCategories[category_id] = true;
+    } else {
+      delete selectedCategories[category_id];
+    }
+    this.setState({ selected_categories: selectedCategories });
+  };
 
-    const rubricTablecolumns = [
-      {
-        name: "rubric_name",
-        label: "Rubric",
-        options: {
-          filter: true,
-          align: "center",
-          customBodyRender: (rubric_name) => {
-            return <p>{rubric_name}</p>;
-          },
-        },
-      },
-      {
-        name: "category_total",
-        label: "Category Total",
-        options: {
-          filter: true,
-          align: "center",
-          customBodyRender: (category_total) => {
-            return <p>{category_total} Categories</p>;
-          },
-        },
-      },
-      {
-        name: "rubric_id",
-        label: "View",
-        options: {
-          filter: false,
-          sort: false,
-          setCellHeaderProps: () => {
-            return {
-              align: "center",
-              width: "100px",
-              className: "button-column-alignment",
-            };
-          },
-          setCellProps: () => {
-            return {
-              align: "center",
-              width: "100px",
-              className: "button-column-alignment",
-            };
-          },
-          customBodyRender: (rubric_id) => {
-            if (rubric_id && categories) {
-              return (
-                <IconButton
-                  id=""
-                  onClick={() => {
-                    this.setState({
-                      chosen_rubric: rubric_id,
-                    });
-                  }}
-                >
-                  <VisibilityIcon sx={{ color: "black" }} />
-                </IconButton>
-              );
-            } else {
-              return <p>{"N/A"}</p>;
-            }
-          },
-        },
-      },
-    ];
-    const categoryTableColumnsAdd = [
-      {
-        name: "category_name",
-        label: "Category",
-        options: {
-          filter: true,
-          align: "center",
-          customBodyRender: (category_name) => {
-            return <p>{category_name}</p>;
-          },
-        },
-      },
-      {
-        name: "rubric_name",
-        label: "Rubric",
-        options: {
-          filter: true,
-          align: "center",
-          customBodyRender: (rubric_name) => {
-            return <p>{rubric_name}</p>;
-          },
-        },
-      },
-      {
-        name: "category_id",
-        label: "Add",
-        options: {
-          filter: false,
-          sort: false,
-          setCellHeaderProps: () => {
-            return {
-              align: "center",
-              width: "100px",
-              className: "button-column-alignment",
-            };
-          },
-          setCellProps: () => {
-            return {
-              align: "center",
-              width: "100px",
-              className: "button-column-alignment",
-            };
-          },
-          customBodyRender: (category_id) => {
-            return (
-              <IconButton
-                id=""
-                onClick={() => {
-                  this.add_category(category_id);
-                }}
-              >
-                <AddCircleIcon sx={{ color: "black" }} />
-              </IconButton>
-            );
-          },
-        },
-      },
-    ];
+  render() {
+    const { rubrics, categories } = this.props;
+    const { selected_categories, category_map } = this.state;
+
+    // NOTE: May be moved outside of the render function
     const categoryTableColumnsRemove = [
       {
         name: "category_name",
@@ -242,7 +190,6 @@ class AddCustomRubric extends React.Component {
         name: "rubric_name",
         label: "Rubric",
         options: {
-          filter: true,
           align: "center",
           customBodyRender: (rubric_name) => {
             return <p>{rubric_name}</p>;
@@ -297,8 +244,6 @@ class AddCustomRubric extends React.Component {
       filter: false,
       viewColumns: false,
     };
-    
-    var categories_of_chosen_rubric = this.state.chosen_rubric === null ? []: categories_by_rubric_id[this.state.chosen_rubric];
 
     var picked_categories = [];
 
@@ -310,37 +255,49 @@ class AddCustomRubric extends React.Component {
       return category_id;
     });
 
-    return(
-      <>
-        { this.state.isLoaded && this.state.errorMessage &&
-          <ErrorMessage
-            errorMessage={this.state.errorMessage}
-          />
-        }
-        <div>
-          <h2
-            style={{
-              borderBottom: "1px solid #D9D9D9",
-              paddingTop: "16px",
-              textAlign: "left",
-              bold: true
-            }}>
-            Customize Your Rubric
-          </h2>
 
+    return (
+      <>
+        {this.state.isLoaded && this.state.errorMessage && (
+          <ErrorMessage errorMessage={this.state.errorMessage} />
+        )}
+        <div>
+          <Grid container spacing={10}>
+            <Grid item xs={6}>
+              <h2
+                style={{
+                  borderBottom: "1px solid #D9D9D9",
+                  paddingTop: "16px",
+                  paddingBottom: "16px",
+                  textAlign: "left",
+                  bold: true,
+                }}
+              >
+                Customize Your Rubric
+              </h2>
+            </Grid>
+            <Grid item xs={6} container justifyContent="flex-end">
+              <CustomButton label="Create Rubric" isOutlined={false} onClick={() => {this.handleCreateRubric(picked_categories)}}/>
+            </Grid>
+          </Grid>
+
+          {/* NOTE: Displays Tables */}
           <Grid container spacing={6.5}>
             <Grid item xs={6}>
               <h3 className="d-flex mb-3">Rubrics</h3>
-              <CustomDataTable
-                data={rubrics ? rubrics : []}
-                columns={rubricTablecolumns}
-                options={options}
+              <CollapsableTableComponent
+                categories={categories}
+                rubrics={rubrics}
+                onCategorySelect={this.handleCategorySelect}
               />
             </Grid>
             <Grid item xs={6}>
-              <h3 className="mb-3">Selected Categories</h3>
+              <h3 className="d-flex mb-3">Selected Categories</h3>
               <CustomDataTable
-                data={categories ? picked_categories : []}
+                data={categories ? Object.keys(selected_categories).map(category_id => ({
+                  category_id,
+                  category_name: category_map[category_id] // Retrieve category name from category_map
+                })) : []}
                 columns={categoryTableColumnsRemove}
                 options={options}
               />
@@ -349,10 +306,7 @@ class AddCustomRubric extends React.Component {
         </div>
       </>
     );
-
   }
-
 }
-
 
 export default AddCustomRubric;
